pyproject.toml
[project]
name = "slides"
version = "2.0.0"
description = "Presentation armory — 7 engines, 30 styles, progressive disclosure. Claude Code plugin for slides, posters, and visual output."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "rooftop-Owl"}]
keywords = ["slides", "presentation", "pptx", "claude-code", "plugin", "design"]

dependencies = [
    "python-pptx>=1.0.0",
]

[project.optional-dependencies]
qa = ["Pillow>=10.0.0"]
browser = ["playwright>=1.40.0"]
all = ["Pillow>=10.0.0", "playwright>=1.40.0"]

[project.urls]
Homepage = "https://github.com/rooftop-Owl/slides"
Repository = "https://github.com/rooftop-Owl/slides"
